Fix jest error
